Add CLI to prompt to select config files
Prompts the user to select the config files, then select the attributes
for each config file. If one of those attributes is `Default`, then the
user is also asked if they want to overwrite or append.

If the user chooses to overwrite, and a "default" was selected, that
config file will be overwritten. Otherwise, it will be appended as
normal.

Resolves #1